Add ERC721 test
